+++ internal/pkg/metrics/combined-registry.go
@@ -36,25 +36,25 @@ func (r *combinedRegistry) SetSSLDaysLeft(job string, daysLeft float64) {
 	}
 }
 
-func (r *combinedRegistry) SetConnectTime(job string, millis int64) {
+func (r *combinedRegistry) SetConnectTime(job string, millis float64) {
 	for _, registry := range r.registries {
 		registry.SetConnectTime(job, millis)
 	}
 }
 
-func (r *combinedRegistry) SetTTFB(job string, millis int64) {
+func (r *combinedRegistry) SetTTFB(job string, millis float64) {
 	for _, registry := range r.registries {
 		registry.SetTTFB(job, millis)
 	}
 }
 
-func (r *combinedRegistry) SetRequestTime(job string, millis int64) {
+func (r *combinedRegistry) SetRequestTime(job string, millis float64) {
 	for _, registry := range r.registries {
 		registry.SetRequestTime(job, millis)
 	}
 }
 
-func (r *combinedRegistry) SetBytesReceived(job string, bytes int64) {
+func (r *combinedRegistry) SetBytesReceived(job string, bytes float64) {
 	for _, registry := range r.registries {
 		registry.SetBytesReceived(job, bytes)
 	}
