+++ internal/pkg/metrics/model.go
@@ -9,10 +9,10 @@ type Registry interface {
 	// SSL Check
 	SetSSLDaysLeft(job string, daysLeft float64)
 	// HTTP Check
-	SetConnectTime(job string, millis int64)
-	SetTTFB(job string, millis int64)
-	SetRequestTime(job string, millis int64)
-	SetBytesReceived(job string, bytes int64)
+	SetConnectTime(job string, millis float64)
+	SetTTFB(job string, millis float64)
+	SetRequestTime(job string, millis float64)
+	SetBytesReceived(job string, bytes float64)
 
 	//DNS lookup check
 	SetIpsRatio(job string, ratio float64)
